Split InMemoryRepository into smaller chunks
